requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
Django==6.0.3
django-htmx==1.27.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-tinymce==5.0.0
idna==3.11
lxml==6.0.2
mail-parser==4.2.1
msal==1.35.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-docx==1.2.0
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
